Quick fix for the google accept cookies page
When you try to search, google will ask if you accept the cookies. As this page is not scrollable it will cancel and throw a "no results found" back.
I added a quick fix in line 60. Its certainly not the best approach but it at least makes the scraper work again.